pyproject.toml
[project]
name = "test-cases-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
    "streamlit>=1.50.0",
]
